requirements.txt
streamlit
langchain-openai
pinecone-client
numpy
pydantic
structlog
langchain